+++ Core/src/main/java/org/eltech/ddm/inputdata/multistreams/MiningMultiStream.java
@@ -1,8 +1,6 @@
-package org.eltech.ddm.inputdata.file.MultiInputStream;
+package org.eltech.ddm.inputdata.multistreams;
 
 import org.eltech.ddm.inputdata.MiningInputStream;
-import org.eltech.ddm.inputdata.MiningVector;
-import org.eltech.ddm.inputdata.file.MiningFileStream;
 import org.eltech.ddm.inputdata.file.csv.ParsingValues;
 import org.eltech.ddm.miningcore.MiningException;
 import org.eltech.ddm.miningcore.miningdata.ELogicalData;
@@ -17,32 +15,18 @@ import java.util.List;
  * @author Maxim Kolpashikov
  */
 
-public abstract class MiningMultiStream extends MiningFileStream {
+public abstract class MiningMultiStream extends MiningInputStream {
 
     protected boolean isOpen = false;
     protected List<ParsingValues> parsingValues;
 
-    protected int vectorsNumber = 0;
     protected MiningInputStream[] streams;
-
-    protected ELogicalData logicalData;
-    protected EPhysicalData physicalData;
 
     // -----------------------------------------------------------------------
     //  Abstract methods
     // -----------------------------------------------------------------------
 
-    public abstract void close() throws MiningException;
-
-    public abstract void open() throws MiningException;
-
-    public abstract void reset() throws MiningException;
-
-    public abstract MiningMultiStream getCopy() throws MiningException;
-
-    public abstract MiningVector next() throws MiningException ;
-
-    public abstract MiningVector getVector(int pos) throws MiningException ;
+    public abstract MiningMultiStream deepCopy() throws MiningException;
 
     // -----------------------------------------------------------------------
     //  Get methods
@@ -50,11 +34,10 @@ public abstract class MiningMultiStream extends MiningFileStream {
 
     /**
      * Returns the number of vectors.
-     *
      * @return the number of vectors
      */
+    @Override
     public int getVectorsNumber() throws MiningException {
-        open();
         return vectorsNumber;
     }
 
@@ -62,17 +45,22 @@ public abstract class MiningMultiStream extends MiningFileStream {
      * Returns a logical data.
      * @return logical data
      */
+    @Override
     public ELogicalData getLogicalData() throws MiningException {
-        open();
         return logicalData;
+    }
+
+    @Override
+    public EPhysicalData recognize() throws MiningException {
+        return physicalData;
     }
 
     /**
      * Returns a physical data.
      * @return physical data
      */
+    @Override
     public EPhysicalData getPhysicalData() throws MiningException {
-        open();
         return physicalData;
     }
 }
